refactor: isolate markdown projection
