+++ plugin/identity.json
@@ -2,5 +2,5 @@
   "schema_version": 1,
   "plugin_version": "0.74.5",
   "hook_manifest_sha256": "ea53d65130d73a4a438d8e9040a17462db52fa3c23e2a0331ed87268543644f4",
-  "inventory_sha256": "0a178c445265c4055fca92e3b0bc05ada5e559953d2292314e01b52a81e40d91"
+  "inventory_sha256": "c3dbbce406396389c1d494a7a0ae62c96f89891bf53c08184587ae5e334458bc"
 }
